Doesn't seem to make it faster

+++ tests/varlen.py
@@ -68,19 +68,17 @@ def assert_close(varname, a, b, eps):
 
 
 class TestClass:
-
     @pytest.mark.parametrize('batch_size', [2, 4])
     @pytest.mark.parametrize('num_heads', [8, 4, 2, 1, 7])
     @pytest.mark.parametrize('head_dim', [64, 32, 16, 50])
     @pytest.mark.parametrize('length', [4096, 2048, 1024, 512, 500])
-    @pytest.mark.parametrize('dtype', [torch.bfloat16])
+    @pytest.mark.parametrize('dtype', [torch.float32])
     def test_varlen(self, batch_size, num_heads, head_dim, length, dtype):
         torch.set_printoptions(linewidth=1024, edgeitems=500)
         device = torch.device('cuda:0')
         lengths = torch.randint(length // 2, length, (batch_size,)).to(device=device, dtype=torch.int32)
         total_length = lengths.sum()
         cu_seqlens = torch.cumsum(lengths, dim=-1)
-
         q = 0.5 * torch.randn((num_heads, total_length, head_dim), device=device, dtype=dtype)
         k = 0.5 * torch.randn((num_heads, total_length, head_dim), device=device, dtype=dtype)
         v = 0.5 * torch.randn((num_heads, total_length, head_dim), device=device, dtype=dtype)
